Switch to a static atomic flag instead of a QSetting
Going through QSettings doesn't work quite well on Windows; setting it
in Quotest and immediately testing through another QSettings instance
inside NetworkAccessManager (in another thread) returns the previous
value. In fact, there's no general guarantee for that, as almost all
QSettings member functions are not thread-safe. With the internal flag,
using direct media requests can be hardcoded to either value inside
a given client, or stored in a QSetting; but the strategy is defined
by the client, which is a good thing too.