ConnectionEncryptionData and per-connection E2EE
connection.cpp has long become a very long read, and E2EE additions
made it even bigger. Those additions, meanwhile, can be rather easily
decoupled and put to a separate class.

On the other hand, it very soon became clear (#603) that a single
build-time switch for E2EE is not enough and there should be a runtime
switch, preferably - per-Connection. This enables the use case where
E2EE is built into the library but clients offer users to opt-in on some
experimental account(s) while keeping the current beta code out of
more sensitive ones, to protect them from accidental breakage.

Introducing such a switch becomes much easier if encryption pieces are
moved out to a separate sub-object - basically, this sub-object is only
created if E2EE is on, and it is technically impossible to call any
operations concerned with E2EE state on an account where it is turned
off.